[backport] Remove the built-in markdown rendering for notebooks (#128806)
This switches our notebooks to always use contributed markdown rendering instead of our built-in markdown renderer

We'd held off on switching over due to accessibility. I've tried to address this using the `aria-describedby` attributed to link the markdown container with a copy of rendered (and sanitized) html from the webview.